The configService's port is now fully configurable
$  ./wrk -t200 -c200 -d10 --latency http://127.0.0.1:1234
Running 10s test @ http://127.0.0.1:1234
  200 threads and 200 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   259.02ms    2.24ms 270.25ms   86.45%
    Req/Sec     3.13      0.49     5.00     93.50%
  Latency Distribution
     50%  259.02ms
     75%  259.67ms
     90%  260.61ms
     99%  268.76ms
  7600 requests in 10.00s, 2.14MB read
Requests/sec:    759.71
Transfer/sec:    218.96KB

$  ./wrk -t200 -c200 -d10 --latency http://127.0.0.1:1024
Running 10s test @ http://127.0.0.1:1024
  200 threads and 200 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   257.39ms    1.96ms 267.34ms   83.86%
    Req/Sec     3.40      0.56     4.00     53.83%
  Latency Distribution
     50%  257.16ms
     75%  258.27ms
     90%  259.20ms
     99%  266.14ms
  7610 requests in 10.01s, 2.13MB read
Requests/sec:    760.57
Transfer/sec:    218.17KB

+++ src/proxy_c/config_service.go
@@ -10,9 +10,9 @@ import (
 	"errors"
 )
 
-func ConfigServer(port int, routeContexts *RoutingContexts) {
+func ConfigServer(port float64, routeContexts *RoutingContexts) {
 	urlRegex := regexp.MustCompile("/server/([a-z0-9-]*){1}")
-	http.ListenAndServe(":"+strconv.Itoa(port), &RegexpHandler{
+	http.ListenAndServe(":"+strconv.Itoa(int(port)), &RegexpHandler{
 		requestMappings: []*requestMapping{
 		&requestMapping{pattern: regexp.MustCompile("/server"), method: "PUT", handler: PUTHandler(func() uuid.UUID {
 			return uuid.NewUUID()
@@ -103,7 +103,6 @@ func GETHandler(urlRegex *regexp.Regexp) func(*RoutingContexts, http.ResponseWri
 		}
 	}
 }
-
 
 
 func DeleteHandler(urlRegex *regexp.Regexp) func(*RoutingContexts, http.ResponseWriter, *http.Request) {
